### 2019-09-13 #1730 @Kevin-Mattheus-Moerman
@arfon `whedon` can't find the paper. Should we instruct the authors to move it to a `paper` folder?
