### 2019-09-15 #1730 @arfon
> @arfon whedon can't find the paper. Should we instruct the authors to move it to a paper folder? The paper is currently here: https://github.com/Chaste/Chaste/blob/release/docs/joss-paper/paper.md

Sometimes Whedon doesn't find the paper on the first try...
