### 2019-09-16 #1730 @meg-simula
Hi @finsberg, would you like to review this submission for JOSS? Feel free to take a look at our reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
