### 2019-09-17 51bec75
Added new neighbour edge operation
