+++ mesh/src/common/EdgeOperation.hpp
@@ -45,7 +45,8 @@ enum EDGE_OPERATION {
     EDGE_OPERATION_DIVIDE,
     EDGE_OPERATION_SPLIT,
     EDGE_OPERATION_NODE_MERGE,
-    EDGE_OPERATION_MERGE
+    EDGE_OPERATION_MERGE,
+    EDGE_OPERATION_NEW_NEIGHBOUR
 };
 
 /**
