### 2019-09-20 #1730 @fcooper8472
@finsberg no problem - I have opened the issues on the Chaste repo.

Probably best as it will save you having to register on our trac.
