### 2019-09-23 1ac0751
[IB] Update docs WRT 2019.1
