+++ docs/licencing/README_LICENCING.md
@@ -1,0 +1,11 @@
+# 3-clause BSD licence
+
+Chaste is distributed as open source software under the [3-clause BSD licence](https://opensource.org/licenses/BSD-3-Clause).
+For full details see the file [Copying.pdf](Copying.pdf).
+
+Chaste uses various third party libraries which have their own licences. 
+For details of these licences and the impact they may have on your use of 
+Chaste please see [Licences.html](Licences.html).
+
+If you contribute to Chaste, your contributions must also be licenced under the 3-clause BSD licence.
+If your employer may claim ownership of your intellectual property, you must ensure that your employer fills out the [Contributor Licence Agreement](ContributorLicenceAgreement.docx) and returns it to the University of Oxford.
