### 2019-09-23 12ee032
#3016 Update copyright and add file to check script
