### 2019-10-14 0d9a912
Added SteadyStateExtrapolationRunner, a class used to find the steady state of action potential models.
TestSteadyStateExtrapolationRunner.hpp provides an example of its usage.
