+++ heart/src/odes/single_cell/SteadyStateExtrapolationRunner.cpp
@@ -1,0 +1,203 @@
+/*
+
+Copyright (c) 2005-2018, University of Oxford.
+All rights reserved.
+
+University of Oxford means the Chancellor, Masters and Scholars of the
+University of Oxford, having an administrative office at Wellington
+Square, Oxford OX1 2JD, UK.
+
+This file is part of Chaste.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice,
+   this list of conditions and the following disclaimer.
+ * Redistributions in binary form must reproduce the above copyright notice,
+   this list of conditions and the following disclaimer in the documentation
+   and/or other materials provided with the distribution.
+ * Neither the name of the University of Oxford nor the names of its
+   contributors may be used to endorse or promote products derived from this
+   software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE
+GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT
+OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+*/
+
+#include "SteadyStateExtrapolationRunner.hpp"
+
+void SteadyStateExtrapolationRunner::RunToSteadyStateImplementation(){
+  mpModel->SetMinimalReset(true);
+  for(mNumEvaluations  = 0; mNumEvaluations < mMaxNumPaces; mNumEvaluations++){
+    if(RunPace() == true){
+      //We have reached the threshold
+      return;
+    }
+  }
+  mpModel->SetMinimalReset(false);
+  return;
+}
+
+
+bool SteadyStateExtrapolationRunner::RunPace(){
+    bool extrapolated = false;
+
+    extrapolated = ExtrapolateStates();
+
+    if(!extrapolated){
+    /*Solve in two parts*/
+      try{
+	mpModel->SolveAndUpdateState(0, mDuration);
+	mpModel->SolveAndUpdateState(mDuration, mPeriod);
+      }
+      catch(Exception &e){
+	std::cout << "RunPace() failed - returning to old state_variables\n";
+	mStateVariables = mSafeStateVariables;
+	mpModel->SetStateVariables(mStateVariables);
+	mMrmsBuffer.clear();
+	mStatesBuffer.clear();
+	return false;
+      }
+      std::vector<double> new_state_variables = mpModel->GetStdVecStateVariables();
+      mStatesBuffer.push_back(new_state_variables);
+      double current_mrms = CalculateMrms(new_state_variables, mStateVariables);
+      mMrmsBuffer.push_back(current_mrms);
[…]

[…]
