### 2019-10-14 #1730 @fcooper8472
From the list, how about one of: dirmeier Rgcjk andremrsantos camillescott sirsharpest mskwark MikeLydeamore

Those, I think, are the list of reviewers who have listed Computational Biology and have C++ as a language preference.
