### 2019-10-15 #1730 @meg-simula
@camillescott would you like to review this submission for JOSS? Feel free to take a look at our reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
