### 2019-10-29 #1730 @meg-simula
@dirmeier @Rgcjk @andremrsantos @sirsharpest @mskwark @MikeLydeamore would any of you review this submission for JOSS? Feel free to take a look at our reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
