### 2019-11-01 38995ef
[JOSS] Ensure consistency between Ubuntu and docker instructions
