### 2019-11-02 #1848 @fcooper8472
@finsberg looks as though the markdown list I put in with the new instructions hasn't rendered nicely.  It's just a standard list in the source - any ideas?
