### 2019-11-02 #1848 @finsberg
I am not sure, but one thing you could try is to add a new line between each item. You can just execute the command `@whedon generate pdf` as a comment here to check the pdf afterwards.
