+++ global/test/TestCitations.hpp
@@ -123,7 +123,9 @@ public:
             PetscSetupUtils::CommonFinalize(); // This prints the citations to disk
 
 // Check PETSc version - this is just because they reformatted their BibTex between versions, no change to function!
-#if (PETSC_VERSION_MAJOR == 3 && (PETSC_VERSION_MINOR == 9 || PETSC_VERSION_MINOR == 10 || PETSC_VERSION_MINOR == 11)) //PETSc 3.9, 3.10 and 3.11
+#if (PETSC_VERSION_MAJOR == 3 && PETSC_VERSION_MINOR == 11) //PETSc 3.11
+            FileFinder reference_citations("global/test/data/citations-2019.txt", RelativeTo::ChasteSourceRoot);
+#elif (PETSC_VERSION_MAJOR == 3 && (PETSC_VERSION_MINOR == 9 || PETSC_VERSION_MINOR == 10)) //PETSc 3.9, 3.10
             FileFinder reference_citations("global/test/data/citations-2018.txt", RelativeTo::ChasteSourceRoot);
 #elif (PETSC_VERSION_MAJOR == 3 && PETSC_VERSION_MINOR == 8) //PETSc 3.8
             FileFinder reference_citations("global/test/data/citations-2017.txt", RelativeTo::ChasteSourceRoot);
