### 2019-11-05 1b63220
#3021 new citation for PETSc >= 3.11.2
