+++ global/src/parallel/PetscTools.cpp
@@ -34,12 +34,12 @@ OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
 #include "PetscTools.hpp"
-#include "Exception.hpp"
-#include "Warnings.hpp"
-#include <iostream>
-#include <sstream>
 #include <cassert>
 #include <cstring> // For strcmp etc. Needed in gcc-4.3
+#include <iostream>
+#include <sstream>
+#include "Exception.hpp"
+#include "Warnings.hpp"
 
 bool PetscTools::mPetscIsInitialised = false;
 unsigned PetscTools::mNumProcessors = 0;
@@ -65,11 +65,11 @@ void PetscTools::ResetCache()
 
         PetscInt num_procs;
         MPI_Comm_size(PETSC_COMM_WORLD, &num_procs);
-        mNumProcessors = (unsigned) num_procs;
+        mNumProcessors = (unsigned)num_procs;
 
         PetscInt my_rank;
         MPI_Comm_rank(PETSC_COMM_WORLD, &my_rank);
-        mRank = (unsigned) my_rank;
+        mRank = (unsigned)my_rank;
     }
     else
     {
@@ -138,11 +138,15 @@ void PetscTools::Barrier(const std::string callerId)
     {
 #ifdef DEBUG_BARRIERS
         // "Before" is alphabetically before "Post" so that one can sort the output on process/event/barrier
-        std::cout << "DEBUG: proc " << PetscTools::GetMyRank() << ": Before " << "Barrier " << mNumBarriers << " \""<< callerId <<  "\"." << std::endl << std::flush;
+        std::cout << "DEBUG: proc " << PetscTools::GetMyRank() << ": Before "
+                  << "Barrier " << mNumBarriers << " \"" << callerId << "\"." << std::endl
+                  << std::flush;
 #endif
         PetscBarrier(PETSC_NULL);
 #ifdef DEBUG_BARRIERS
-        std::cout << "DEBUG: proc " << PetscTools::GetMyRank() << ": Post " << "Barrier " << mNumBarriers++ << " \""<< callerId <<  "\"." << std::endl << std::flush;
+        std::cout << "DEBUG: proc " << PetscTools::GetMyRank() << ": Post "
+                  << "Barrier " << mNumBarriers++ << " \"" << callerId << "\"." << std::endl
+                  << std::flush;
 #endif
     }
 }
@@ -151,7 +155,7 @@ void PetscTools::BeginRoundRobin()
 {
     Barrier("PetscTools::RoundRobin"); // We want barriers both before all and after all, just in case
     const unsigned my_rank = GetMyRank();
-    for (unsigned turn=0; turn<my_rank; turn++)
+    for (unsigned turn = 0; turn < my_rank; turn++)
     {
         Barrier("PetscTools::RoundRobin");
     }
@@ -160,7 +164,7 @@ void PetscTools::BeginRoundRobin()
 void PetscTools::EndRoundRobin()
 {
     const unsigned num_procs = GetNumProcs();
-    for (unsigned turn=GetMyRank(); turn<num_procs; turn++)
+    for (unsigned turn = GetMyRank(); turn < num_procs; turn++)
     {
         Barrier("PetscTools::RoundRobin");
     }
@@ -186,7 +190,7 @@ MPI_Comm PetscTools::GetWorld()
 bool PetscTools::ReplicateBool(bool flag)
 {
     CheckCache();
-    unsigned my_flag = (unsigned) flag;
+    unsigned my_flag = (unsigned)flag;
     unsigned anyones_flag_is_true = my_flag;
[…]
