### 2019-11-08 #1848 @SirSharpest
I have started to review, but from this submission I cannot see what specifically is different from https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1002970 or https://www.sciencedirect.com/science/article/pii/S0010465509002604.

If it is a specific recent addition to this existing and established software, then could the submission be made to reflect this?

Otherwise I would need @meg-simula to verify if established and published software is an issue / needs to be taken into account in this review?
