### 2019-11-08 #1848 @fcooper8472
Hi @SirSharpest.  The paper you mention was published in 2009, and there has been a vast quantity of development since then.

This paper references that publication, and another (from 2013), but this will serve as an updated reference with details about Chaste as it exists now (rather than 6 or 10 years ago).

Almost every element of the software has been substantially changed, and information on the installation and example usage reflects the current release of the software:

- All three of the examples of functionality relate to new functionality since the 2009 paper you reference
- Information on how to run the examples reflects a completely new build system
- Installation instructions reflect new functionality including Docker

As this is a software paper (which will allow the citing of the software) I think it would not be useful to highlight how the current version is different from previous versions, but simply (as is presented) what the current version of the software offers.

Certainly happy to hear any input from @meg-simula on this though.
