### 2019-11-08 #1848 @finsberg
@meg-simula I am now done with my review, and I think that Chaste more than qualifies for publication in JOSS. Its seems to have an active community of maintainers and it also has a very nice test suite as well a good documentation with a lot of tutorials. It was also very easy to use together with the provided docker image. I have added a final suggestion to add a section for benchmarks which I think would be a very nice addition.
