### 2019-11-12 #1848 @meg-simula
> Otherwise I would need @meg-simula to verify if established and published software is an issue / needs to be taken into account in this review?

Thanks for the due diligence @SirSharpest! In this case, the difference between the aforementioned published papers and the current submission, in terms of target audience, scope and new development seems substantial and warrants review. But please do feel free to add a pointer in your review if you would like this point to be addressed more clearly in the manuscript.
