### 2019-11-12 #1848 @meg-simula
Thanks for the reviewing @finsberg! There is one checkbox (state of the art) unchecked in your list - is that on purpose?
