### 2019-11-12 #1848 @finsberg
> Thanks for the reviewing @finsberg! There is one checkbox (state of the art) unchecked in your list - is that on purpose?

There is a reason for this. The question in the checkbox says " Do the authors describe how this software compares to other commonly-used packages?" and I cannot see that they have any comparison with other commonly-used packages. Thats said, I think adding a section with benchmarks would be a good solution to address this question. However, in my opinion I think Chaste qualifies for publication also without this benchmark section.
