### 2019-11-12 #1848 @meg-simula
Ok, thanks for the clarification. I would then advise @fcooper8472 to add a discussion on comparison with other packages (in addition to the optional benchmark comparison).
