### 2019-11-19 #1848 @SirSharpest
@meg-simula Could I ask to be removed as a reviewer? I have ethical concerns over this which make it so that I cannot be fully objective.
