### 2019-11-19 #1848 @mirams
It would be good to get some clarification on this (I’m one of the authors of the Chaste paper).

As I understood it, the central motivation for a JOSS paper is to give all the authors that have contributed to a certain version of an open source scientific code some credit and citations for their work, and to provide a short summary of the capabilities and science the software has enabled.

We’ve listed all the authors that have committed more than once or twice to the git repository (also some svn before that) since the last (2013) paper. We would be planning to submit another new updated paper in a year or two with a new set of authors who have contributed code from now until then, if this isn’t allowed or encouraged by JOSS that would be useful to know.
