### 2019-11-19 #1848 @meg-simula
I will discuss the matter with the JOSS editorial board and get back to you all.
