### 2019-11-27 #1848 @meg-simula
I've consulted with the editorial board and the main points are as follows:

(1) As the previous papers are published in Software "tracks" e.g. PLOS Computational Biology: Software and a special issue of Computer Physics Communications focused on high quality software and indeed focus on software aspects of the project, these papers do indeed document the Chaste software.

(2) A JOSS paper is admissible if and only if the software has substantial new additions since the previous papers.

Thus, if it is the case as @fcooper8472 states that "Almost every element of the software has been substantially changed", I would consider this to constitute substantial new additions and thus that this paper is admissible. I would however encourage the authors to present the main additions compared to the previous versions in the current paper.

@mirams: The same thus applies for future versions - substantial new additions since previous papers required. Hope this clarifies, please let me know if not.

@SirSharpest: would you still like to be removed as reviewer? If so, no problem, just let me know.
