### 2019-12-04 3fb0226
Node pointers in Edge class are now updated after node reindexing
