### 2019-12-10 b71ac52
VertexMeshOperationRecorder added
