+++ cell_based/src/writers/population_writers/VertexT1SwapLocationsWriter.cpp
@@ -71,7 +71,6 @@ template<unsigned ELEMENT_DIM, unsigned SPACE_DIM>
 void VertexT1SwapLocationsWriter<ELEMENT_DIM, SPACE_DIM>::Visit(VertexBasedCellPopulation<SPACE_DIM>* pCellPopulation)
 {
     std::vector< c_vector<double, SPACE_DIM> > t1_swap_locations = pCellPopulation->rGetMesh().GetLocationsOfT1Swaps();
-
     *this->mpOutStream << t1_swap_locations.size() << "\t";
     for (unsigned index = 0;  index < t1_swap_locations.size(); index++)
     {
