### 2019-12-11 0cd9507
VertexMeshOperationRecorder helper class is implemented
The helper class keeps the record of mesh changes (in particular edges) instead of MutableVertexMesh class
