### 2019-12-13 bf9fa36
passing annotate=False flag when adding cytosolic_calcium_variable as output to prevent accidentally annotating it
