### 2019-12-16 563f3ee
Element deletion is registered with EdgeHelper class
