### 2019-12-16 a5a80c7
Element deletion is registered with EdgeHelper class
