### 2019-12-21 #1848 @labarba
@fcooper8472 — it looks like the handling editor is waiting for a reply from you. Can you update this thread with your assessment and how you wish this to move forward?
