### 2020-01-10 d6f4dd3
Fixed various bugs and added support for previously unaccounted events
T2 swaps are now properly taken into account and cell srns are updated properly. 
Fixed several bugs regarding node merging and their effect on edge SRNs.
Added support for the case when triangular cells divide. 
Added some documentation
