+++ cell_based/src/population/srn_update/VertexBasedPopulationSrn.cpp
@@ -45,50 +45,59 @@ void VertexBasedPopulationSrn<DIM>::SetVertexCellPopulation(VertexBasedCellPopul
 }
 
 template <unsigned DIM>
-void VertexBasedPopulationSrn<DIM>::UpdateSrnAfterBirthOrDeath()
+void VertexBasedPopulationSrn<DIM>::UpdateSrnAfterBirthOrDeath(VertexElementMap& rElementMap)
 {
     std::vector<EdgeOperation*> edge_operations
     = mpCellPopulation->rGetMesh().GetOperationRecorder()->GetEdgeOperations();
-
     for (auto operation:edge_operations)
     {
-        switch (operation->GetOperation())
+        // An operation with deleted element may be recorded,
+        // e.g. following a T2 swap, neighbouring element may become triangular due to node merging
+        // and may subsequently be marked for deletion, despite being marked to perform edge operations.
+        if (!rElementMap.IsDeleted(operation->GetElementIndex()))
         {
-        case EDGE_OPERATION_ADD:
-        case EDGE_OPERATION_NODE_MERGE:
-        case EDGE_OPERATION_SPLIT:
-        case EDGE_OPERATION_MERGE:
-        case EDGE_OPERATION_NEW_NEIGHBOUR:
-        {
-            const unsigned int location_index = operation->GetElementIndex();
-            EdgeRemapInfo *pEdgeChange = operation->GetNewEdges();
-            CellPtr cell = mpCellPopulation->GetCellUsingLocationIndex(location_index);
-            auto old_model = static_cast<SrnCellModel*>(cell->GetSrnModel());
-            std::vector<AbstractSrnModelPtr> old_srn_edges = old_model->GetEdges();
-            RemapCellSrn(old_srn_edges, old_model, pEdgeChange);
-            old_srn_edges.clear();
-            break;
-        }
-        case EDGE_OPERATION_DELETE:
-            break;
-        case EDGE_OPERATION_DIVIDE:
-        {
-            const unsigned int location_index_1 = operation->GetElementIndex();
-            const unsigned int location_index_2 = operation->GetElementIndex2();
-            EdgeRemapInfo *pEdgeChange_1 = operation->GetNewEdges();
-            EdgeRemapInfo *pEdgeChange_2 = operation->GetNewEdges2();
-            CellPtr cell_1 = mpCellPopulation->GetCellUsingLocationIndex(location_index_1);
-            CellPtr cell_2 = mpCellPopulation->GetCellUsingLocationIndex(location_index_2);
+            switch (operation->GetOperation())
+            {
+            case EDGE_OPERATION_ADD:
+            case EDGE_OPERATION_NODE_MERGE:
+            case EDGE_OPERATION_SPLIT:
+            case EDGE_OPERATION_MERGE:
+            {
+                unsigned int location_index = operation->GetElementIndex();
+                if (operation->IsElementIndexRemapped())
+                {
+                    location_index = rElementMap.GetNewIndex(operation->GetElementIndex());
+                }
+                EdgeRemapInfo *pEdgeChange = operation->GetRemapInfo();
+                CellPtr cell = mpCellPopulation->GetCellUsingLocationIndex(location_index);
 
-            auto old_model_1 = static_cast<SrnCellModel*>(cell_1->GetSrnModel());
-            std::vector<AbstractSrnModelPtr> parent_srn_edges = old_model_1->GetEdges();
-            auto old_model_2 = static_cast<SrnCellModel*>(cell_2->GetSrnModel());
+                auto old_model = static_cast<SrnCellModel*>(cell->GetSrnModel());
+                std::vector<AbstractSrnModelPtr> old_srn_edges = old_model->GetEdges();
[…]
