### 2020-01-14 bc3160c
VertexMeshOperationRecorder now stores info about T1-3 swaps for output
T1-3 writers access information from VertexMeshOperationRecorder
If additional information about mesh operations is required, T1-3SwapInfo and CellDivisionInfo structs can be modified.
Redundant methods in MutableVertexMesh and VertexBasedCellPopulation classes are removed
