+++ cell_based/src/writers/population_writers/VertexT1SwapLocationsWriter.cpp
@@ -70,18 +70,18 @@ void VertexT1SwapLocationsWriter<ELEMENT_DIM, SPACE_DIM>::Visit(PottsBasedCellPo
 template<unsigned ELEMENT_DIM, unsigned SPACE_DIM>
 void VertexT1SwapLocationsWriter<ELEMENT_DIM, SPACE_DIM>::Visit(VertexBasedCellPopulation<SPACE_DIM>* pCellPopulation)
 {
-    std::vector< c_vector<double, SPACE_DIM> > t1_swap_locations
-    = pCellPopulation->rGetMesh().GetOperationRecorder()->GetLocationsOfT1Swaps();
-    *this->mpOutStream << t1_swap_locations.size() << "\t";
-    for (unsigned index = 0;  index < t1_swap_locations.size(); index++)
+    std::vector<T1SwapInfo<SPACE_DIM> > t1_swap_info
+    = pCellPopulation->rGetMesh().GetOperationRecorder()->GetT1SwapsInfo();
+    *this->mpOutStream << t1_swap_info.size() << "\t";
+    for (unsigned index = 0;  index < t1_swap_info.size(); index++)
     {
         for (unsigned i=0; i<SPACE_DIM; i++)
         {
-            *this->mpOutStream << t1_swap_locations[index][i] << "\t";
+            *this->mpOutStream << t1_swap_info[index].mLocation[i] << "\t";
         }
     }
 
-    pCellPopulation->rGetMesh().GetOperationRecorder()->ClearLocationsOfT1Swaps();
+    pCellPopulation->rGetMesh().GetOperationRecorder()->ClearT1SwapsInfo();
 }
 
 // Explicit instantiation
