+++ python/infra/CheckForCopyrights.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-"""Copyright (c) 2005-2019, University of Oxford.
+"""Copyright (c) 2005-2020, University of Oxford.
 All rights reserved.
 
 University of Oxford means the Chancellor, Masters and Scholars of the
@@ -70,7 +70,7 @@ deprecated_notice = re.compile(r"""(# ){0,1}Copyright \(c\) 2005-\d{4}, Universi
 """, re.MULTILINE)
 
 
-current_notice="""Copyright (c) 2005-2019, University of Oxford.
+current_notice = """Copyright (c) 2005-2020, University of Oxford.
 All rights reserved.
 
 University of Oxford means the Chancellor, Masters and Scholars of the
@@ -103,27 +103,29 @@ OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 """
 
 license_current_notice = 'BSD 3-Clause License.\n\n'+current_notice
-py_current_notice='"""'+current_notice+'"""\n'
-cpp_current_notice='/*\n\n'+current_notice+'\n*/'
+py_current_notice = '"""'+current_notice+'"""\n'
+cpp_current_notice = '/*\n\n'+current_notice+'\n*/'
+
 
 def AddHashesToBeginningOfLines(message):
     message = '# ' + message.replace("\n", "\n# ")
     return message[:-2]
 
+
 cmake_current_notice = AddHashesToBeginningOfLines(current_notice)
 
-output_notice=current_notice.replace("\nThis file is part of Chaste.\n", "")
+output_notice = current_notice.replace("\nThis file is part of Chaste.\n", "")
 boost_random_distribution_notice = """
  * Distributed under the Boost Software License, Version 1.0. (See
  * accompanying file LICENSE_1_0.txt or copy at
  * http://www.boost.org/LICENSE_1_0.txt)
  """
-pycml_notice=" Processed by pycml - CellML Tools in Python"
-xsd2_notice="// Copyright (C) 2005-2007 Code Synthesis Tools CC"
-xsd3_notice="// Copyright (C) 2005-2008 Code Synthesis Tools CC"
-triangle_notice="""/*  Copyright 1993, 1995, 1997, 1998, 2002, 2005                             */
+pycml_notice = " Processed by pycml - CellML Tools in Python"
+xsd2_notice = "// Copyright (C) 2005-2007 Code Synthesis Tools CC"
+xsd3_notice = "// Copyright (C) 2005-2008 Code Synthesis Tools CC"
+triangle_notice = """/*  Copyright 1993, 1995, 1997, 1998, 2002, 2005                             */
 /*  Jonathan Richard Shewchuk                                                */"""
-tetgen_notice="""///////////////////////////////////////////////////////////////////////////////
+tetgen_notice = """///////////////////////////////////////////////////////////////////////////////
 //                                                                           //
 // TetGen                                                                    //
 //                                                                           //
@@ -145,7 +147,7 @@ tetgen_notice="""///////////////////////////////////////////////////////////////
 //                                                                           //
 ///////////////////////////////////////////////////////////////////////////////
 """
-tetgen_predicates_notice="""/*****************************************************************************/
+tetgen_predicates_notice = """/*****************************************************************************/
 /*                                                                           */
 /*  Routines for Arbitrary Precision Floating-point Arithmetic               */
 /*  and Fast Robust Geometric Predicates                                     */
[…]
