+++ python/hostconfig/machines/oerc.py
@@ -1,6 +1,6 @@
 # Configuration for HAL or SAL
 
-"""Copyright (c) 2005-2019, University of Oxford.
+"""Copyright (c) 2005-2020, University of Oxford.
 All rights reserved.
 
 University of Oxford means the Chancellor, Masters and Scholars of the
