+++ python/hostconfig/machines/archer.py
@@ -1,4 +1,4 @@
-"""Copyright (c) 2005-2019, University of Oxford.
+"""Copyright (c) 2005-2020, University of Oxford.
 All rights reserved.
 
 University of Oxford means the Chancellor, Masters and Scholars of the
