+++ python/pycml/translators.py
@@ -1,6 +1,6 @@
 # We want 1/2==0.5
 from __future__ import division
-"""Copyright (c) 2005-2019, University of Oxford.
+"""Copyright (c) 2005-2020, University of Oxford.
 All rights reserved.
 
 University of Oxford means the Chancellor, Masters and Scholars of the
