### 2020-01-16 c4b397d
Move metadata to its own folder in preparation for swapping in git submodule
