+++ python/pycml/cellml_metadata.py
@@ -1,4 +1,10 @@
 
+import rdflib
+from cStringIO import StringIO
+import types
+import sys
+import os
+import logging
 """Copyright (c) 2005-2020, University of Oxford.
 All rights reserved.
 
@@ -42,14 +48,8 @@ of terms in the ontology that can annotate variables, and the subset of those na
 which define properties of the stimulus current (but not the current itself), respectively.
 """
 
-import logging
-import os
-import sys
-import types
-from cStringIO import StringIO
 
 # We now only support rdflib for RDF processing
-import rdflib
 
 
 def __init__(module):
@@ -60,6 +60,7 @@ def __init__(module):
 
 class RdfProcessor(object):
     """Implements CellML metadata functionality using the RDFLib library."""
+
     def __init__(self, name):
         """Create the wrapper."""
         # Magic for pretending to be a module
@@ -98,8 +99,9 @@ class RdfProcessor(object):
     def _load_ontology(self):
         """Load the Oxford metadata ontology the first time it's needed."""
         pycml_path = os.path.dirname(os.path.realpath(__file__))
-        oxmeta_ttl = os.path.join(pycml_path, 'oxford-metadata.ttl')
-        oxmeta_rdf = os.path.join(pycml_path, 'oxford-metadata.rdf')
+        metadata_path = os.path.join(pycml_path, 'metadata_repo')
+        oxmeta_ttl = os.path.join(metadata_path, 'oxford-metadata.ttl')
+        oxmeta_rdf = os.path.join(metadata_path, 'oxford-metadata.rdf')
 
         g = self._ontology = self.Graph()
         # We allow a difference in modification time of 10s, so we don't get confused when checking out!
@@ -115,12 +117,15 @@ class RdfProcessor(object):
         else:
             # Just parse the RDF/XML version
             g.parse(oxmeta_rdf, format='xml')
-        
-        annotation_terms = list(g.subjects(rdflib.RDF.type, rdflib.URIRef(pycml.NSS['oxmeta']+u'Annotation')))
-        self._metadata_names = frozenset(map(lambda node: self.namespace_member(node, pycml.NSS['oxmeta']), annotation_terms))
-        
+
+        annotation_terms = list(g.subjects(
+            rdflib.RDF.type, rdflib.URIRef(pycml.NSS['oxmeta']+u'Annotation')))
+        self._metadata_names = frozenset(map(lambda node: self.namespace_member(
+            node, pycml.NSS['oxmeta']), annotation_terms))
+
         # Parameters for the stimulus current
-        self._stimulus_names = frozenset(filter(lambda name: name.startswith('membrane_stimulus_current_'), self._metadata_names))
+        self._stimulus_names = frozenset(filter(lambda name: name.startswith(
+            'membrane_stimulus_current_'), self._metadata_names))
 
     @property
     def METADATA_NAMES(self):
@@ -141,7 +146,7 @@ class RdfProcessor(object):
         The new store will be available as self._models[cellml_model].
         """
         self._models[cellml_model] = self.Graph()
-    
+
     def _add_rdf_element(self, cellml_model, rdf_text):
         """Add statements to the model's graph from the given serialized RDF."""
         g = self.Graph()
@@ -149,14 +154,14 @@ class RdfProcessor(object):
         rdf_model = self._models[cellml_model]
         for stmt in g:
             rdf_model.add(stmt)
-    
+
     def _serialize(self, cellml_model):
         """Serialize the RDF model for this CellML model to XML."""
         return self._models[cellml_model].serialize()
-    
+
     def get_rdf_from_model(self, cellml_model):
         """Get the RDF graph of the given CellML model.
-    
+
         If this model is already in our map, return the existing RDF store.
[…]
