### 2020-01-16 b9b8bcd
Fixed serialization problem with VertexBasedCellPopulation class
The problem was due to improper serialization of CellEdgeData class
