+++ mesh/src/vertex/MutableVertexMesh.hpp
@@ -342,7 +342,6 @@ protected:
         archive & mDeletedElementIndices;
         archive & mDistanceForT3SwapChecking;
         ///\todo: maybe we should archive the mLocationsOfT1Swaps and mDeletedNodeIndices etc. as well?
-
         archive & boost::serialization::base_object<VertexMesh<ELEMENT_DIM, SPACE_DIM> >(*this);
     }
 public:
