### 2020-01-21 6ccca8d
Add a new test for CellML metadata submodule
