### 2020-01-21 9c07cbf
Add a nicer PyCML error message if ontology is missing
