### 2020-01-21 f2ab4ef
#3024 add a test to say if the submodule has gone stale
