### 2020-01-21 38e01cc
#3024 Tidy up messages
