### 2020-01-21 25efbd0
#3024 add copyright ignore for submodule python script
