### 2020-01-22 d1f088a
#3025 first attempt at GitHub action
