+++ .github/workflows/ubuntu.yml
@@ -1,0 +1,32 @@
+name: C/C++ CI
+
+on:
+  push:
+    branches:
+      - develop
+
+jobs:
+  
+  build-and-test:
+    name: Test on Ubuntu 18.04
+    runs-on: ubuntu-18.04
+    
+    steps:
+    - uses: actions/checkout@v1
+    - name: install dependencies
+      run: |
+        echo 'deb http://www.cs.ox.ac.uk/chaste/ubuntu bionic/' | sudo tee -a /etc/apt/sources.list.d/chaste.list
+        sudo apt-key adv --recv-keys --keyserver hkp://keyserver.ubuntu.com:80 422C4D99
+        sudo apt update
+        sudo apt install chaste-dependencies
+    - name: make build directory
+      run: mkdir chaste-build-dir
+    - name: cmake configure
+      run: cmake ..
+      working-directory: chaste-build-dir
+    - name: cmake build
+      run: cmake --build . --parallel 2 --target Continuous
+      working-directory: chaste-build-dir
+    - name: cmake test
+      run: ctest -L Continuous -j2
+      working-directory: chaste-build-dir
