### 2020-01-22 d34b12f
#3025 use github instructions for checking out submodule
