### 2020-01-22 2c17530
#3025 Simplify checkout of submodule...
