### 2020-01-22 259bf3c
#3025 output test failures to console
