+++ .github/workflows/ubuntu.yml
@@ -1,4 +1,4 @@
-name: C/C++ CI
+name: Ubuntu continuous test suite
 
 on:
   push:
@@ -30,5 +30,5 @@ jobs:
       run: cmake --build . --parallel 2 --target Continuous
       working-directory: chaste-build-dir
     - name: cmake test
-      run: ctest -L Continuous -j2
+      run: ctest -j2 -L Continuous --output-on-failure
       working-directory: chaste-build-dir
