+++ .github/workflows/macos.yml
@@ -18,44 +18,53 @@ jobs:
     - name: install boost
       run: |
         brew install boost
+      if: always() 
     - name: install xerces
       run: |
         brew install xerces-c
+      if: always() 
     - name: install metis
       run: |
         brew install metis
-    - name: install parmetis
-      run: |
-        brew install parmetis
+      if: always() 
     - name: install sundials
       run: |
         brew install sundials --with-mpi
+      if: always() 
     - name: install hdf5
       run: |
         brew install hdf5 --with-mpi
+      if: always() 
     - name: install vtk
       run: |
         brew install vtk
+      if: always() 
     - name: install petsc
       run: |
         brew install petsc --with-mpi
+      if: always() 
     - name: install vtk
       run: |
         brew install vtk
+      if: always() 
     - name: install xsd
       run: |
         wget https://www.codesynthesis.com/download/xsd/4.0/macosx/i686/xsd-4.0.0-i686-macosx.tar.bz2
         tar -jxf xsd-4.0.0-i686-macosx.tar.bz2
-        brew install vtk
+        ls -ltr
+      working-directory: /usr/local/Cellar/
+      if: always() 
     - name: install PyCML
       run: |
         python2 -m pip install "python-dateutil==1.5"
         python2 -m pip install "rdflib==2.4.2"
         python2 -m pip install lxml
         python2 -m pip install "Amara==1.2.0.2"
+      if: always() 
     - name: install Amara
       run: |
         sudo easy_install https://pypi.python.org/packages/2.5/A/Amara/Amara-1.2.0.2-py2.5.egg
+      if: always() 
     - name: make build directory
       run: mkdir chaste-build-dir
     - name: cmake configure
