### 2020-01-22 580d01e
#3025 Turn off Ubuntu testing while messing with macOS
