### 2020-01-22 76c24fd
#3025 Try without specific --with-mpi options
