### 2020-01-22 ff1765d
Always report CMake version to help with debugging
