### 2020-01-22 416cde8
#3025 Try to work around boost/cmake issue
https://gitlab.kitware.com/cmake/cmake/issues/18865
