### 2020-01-24 #1848 @IgorBaratta
Hi @meg-simula, I have completed my review.

Yet, there are still two boxes unchecked, both concerning the software paper:
- State of the field – As [previosly noted](https://github.com/openjournals/joss-reviews/issues/1848#issuecomment-553086261) by @finsberg, a comparison to other commonly-used packages is missing.
- References – There seem to be some inconsistencies in the citation format in the section “Recent publications enabled by Chaste”.

But in any case, I recommend this submission for publication in JOSS.

The software is exceptionally documented and is a great resource. I was able to follow the tutorials of the core functionality smoothly. Also tried with success the three tutorials presented in the text: cardiac, cell-based, and lung Chaste.

_Minor Remark:_
Installation went well following the “Users of the Ubuntu package” tutorial in a brand new Ubuntu installation. However, I had some issues with a previously installed PETSc that complied with the  Dependency Version guide.
It didn’t work with PETSc compiled with complex numbers. So, I think that this incompatibility, if confirmed, could be mentioned in the documentation in https://chaste.cs.ox.ac.uk/trac/wiki/InstallGuides/DependencyVersions or https://chaste.cs.ox.ac.uk/trac/wiki/InstallPetscAndMpi.
