### 2020-02-05 9d642d6
Added tests for ReMesh operations and edge SRNs
