### 2020-02-10 #1848 @fcooper8472
Sorry for the delay on my end - I'll have a revised version that addresses the two outstanding points shortly.
