### 2020-02-17 1a8c261
[JOSS] Paper revisions
