+++ docs/joss-paper/paper.bib
@@ -203,8 +203,6 @@
   publisher={Public Library of Science}
 }
 
-
-
 @Article{Britton2013Experimentally,
   author  = {Britton, O.J. and Bueno-Orovio, A. and Van Ammel, K. and Lu, H.R. and Towart, R. and Gallacher, D.J. and Rodriguez, B.},
   title   = {{Experimentally calibrated population of models predicts and explains intersubject variability in cardiac cellular electrophysiology}},
@@ -1974,7 +1972,7 @@ MRREVIEWER = {Ali R. Soheili},
        DOI = {10.1137/17M1137061},
        URL = {https://doi-org.cyber.usask.ca/10.1137/17M1137061},
 }
-		
+
 @article {Spiteri2016Godunov,
     AUTHOR = {Spiteri, Raymond J. and Torabi Ziaratgahi, Saeed},
      TITLE = {Operator splitting for the bidomain model revisited},
@@ -2007,4 +2005,120 @@ MRREVIEWER = {Manfred Dobrowolski},
 MRREVIEWER = {},
        DOI = {},
        URL = {},
+}
+
+@article{Swat2012CompuCell3D,
+  author = {Swat, Maciej H and Thomas, Gilberto L and Belmonte, Julio M and Shirinifard, Abbas and Hmeljak, Dimitrij and Glazier, James A},
+  doi = {10.1016/B978-0-12-388403-9.00013-8},
+  journal = {Methods Cell Biol.},
+  pages = {325--366},
+  pmid = {22482955},
+  title = {{Multi-scale modeling of tissues using CompuCell3D}},
+  url = {http://www.ncbi.nlm.nih.gov/pubmed/22482955},
+  volume = {110},
+  year = {2012}
+}
+
+@article{Starruss2014Morpheus,
+  author = {Starru{\ss}, J{\"{o}}rn and de Back, Walter and Brusch, Lutz and Deutsch, Andreas},
+  doi = {10.1093/bioinformatics/btt772},
+  issn = {1460-2059},
+  journal = {Bioinformatics},
+  month = {may},
+  number = {9},
+  pages = {1331--1332},
+  publisher = {Oxford University Press},
+  title = {{Morpheus: a user-friendly modeling environment for multiscale and multicellular systems biology}},
+  url = {https://academic.oup.com/bioinformatics/article-lookup/doi/10.1093/bioinformatics/btt772},
+  volume = {30},
+  year = {2014}
+}
+
+@article{Sutterlin2013EPISIM,
+  author = {Sutterlin, T. and Kolb, C. and Dickhaus, H. and Jager, D. and Grabe, N.},
+  doi = {10.1093/bioinformatics/bts659},
+  issn = {1367-4803},
+  journal = {Bioinformatics},
+  month = {jan},
+  number = {2},
+  pages = {223--229},
+  publisher = {Oxford University Press},
+  title = {{Bridging the scales: semantic integration of quantitative SBML in graphical multi-cellular models and simulations with EPISIM and COPASI}},
+  url = {https://academic.oup.com/bioinformatics/article-lookup/doi/10.1093/bioinformatics/bts659},
+  volume = {29},
+  year = {2013}
+}
+
+@article{Hoehme2010CellSys,
+  author = {Hoehme, Stefan and Drasdo, Dirk},
+  doi = {10.1093/bioinformatics/btq437},
+  journal = {Bioinformatics},
+  month = {oct},
+  number = {20},
+  pages = {2641--2},
+  pmid = {20709692},
+  publisher = {Oxford University Press},
+  title = {{A cell-based simulation software for multi-cellular systems}},
+  url = {http://www.ncbi.nlm.nih.gov/pubmed/20709692},
+  volume = {26},
[…]
