+++ docs/joss-paper/paper.md
@@ -145,12 +145,12 @@ It aims to couple biophysically-detailed models of airway mechanics with organ-s
 
 Chaste is designed to be modular and extensible, providing libraries for common scientific computing infrastructure such as linear algebra operations, finite element meshes, and ordinary and partial differential equation solvers.
 This infrastructure is used by libraries for specific applications, such as continuum mechanics, cardiac models, and cell-based models.
-The software engineering techniques used to develop Chaste are intended to ensure code quality, re-usablity and reliability.
+The software engineering techniques used to develop Chaste are intended to ensure code quality, re-usability and reliability.
 Primary applications of the software include cardiac and respiratory physiology, cancer and developmental biology.
 
 ## The software
 
-Chaste is available on GitHub (https://github.com/Chaste/Chaste), and the current stable release is version 2019.1.
+Chaste is available on GitHub [https://github.com/Chaste/Chaste](https://github.com/Chaste/Chaste), and the current stable release is version 2019.1.
 Please see the Readme.md file on the Github repository for links to the Chaste wiki and install guides.
 
 Previous publications about Chaste have detailed the rationale for, and design principles behind, Chaste [@Pitt-Francis2009Chaste], as well as the main application areas of Chaste up to 2013 [@Mirams2013Chaste].
@@ -159,39 +159,77 @@ Chaste places an emphasis on reproducibility and verification and, as such, exte
 A series of test suites must all pass before any commit is considered a release-candidate.
 Most testing is performed on Long Term Support (LTS) versions of Ubuntu Linux, with unit tests additionally being run on macOS.
 
-Testing includes compilation of all libraries with GCC, clang and Intel C++ compilers; extensive unit testing; performance profiling to identify any slowdowns over time; memory testing with valgrind; verification of code coverage; and running unit tests with different combinations of dependencies to ensure portability.
-The output of these tests is available at
-https://chaste.cs.ox.ac.uk/buildbot/
+Testing includes compilation of all libraries with GCC, Clang and Intel C++ compilers; extensive unit testing; performance profiling to identify any slowdowns over time; memory testing with valgrind; verification of code coverage; and running unit tests with different combinations of dependencies to ensure portability.
+The output of these tests is available at [https://chaste.cs.ox.ac.uk/buildbot/](https://chaste.cs.ox.ac.uk/buildbot/).
+
+Since 2013, Chaste has substantially changed to modernise its infrastructure and to enable new science.
+In terms of infrastructure, Chaste now uses a modern CMake build system, the C++14 language standard, and makes extensive use of BuildBot for continuous integration.
+In terms of science, Lung Chaste is entirely new and allows the use of Chaste in a new scientific domain.
+In Cardiac Chaste, we can now create algebraic Jacobians for CellML ODE systems, which can improve speed of simulation for cardiac action potential and tissue simulations [@Cooper2015Cellular], and metadata annotations of CellML files have replaced manual specification of variables in configuration files.
+Cell-based Chaste has been overhauled to improve flexibility.
+Changes include hierarchies of simulation modifiers, information writers, cell-cycle models, subcellular reaction network models, and numerical methods that allow new customisation points in almost every area of all cell-based simulations.
+In addition, simulation output has been standardised to use VTK, a standard and powerful visualisation framework, and some cell-centre simulations now run in parallel using MPI.
+
+### Comparison with other software
+
[…]
