### 2020-02-17 #1848 @fcooper8472
@IgorBaratta @finsberg @meg-simula please see the latest version, which addresses the following points:

1. Information on the extent of changes since 2013, when the most recent previous paper about Chaste was published
2. A new section with comparison to similar software

Two things I see when looking at the proof:

1. The table (in the new comparison section) is badly formatted due to lengthy references.
2. All cases with multiple refs in one place are not formatted correctly.  The syntax `[@ref1, @ref2]` does not seem to work.

Any ideas on these formatting issues?
