### 2020-02-17 #1848 @finsberg
> @IgorBaratta @finsberg @meg-simula please see the latest version, which addresses the following points:
>
> 1. Information on the extent of changes since 2013, when the most recent previous paper about Chaste was published
> 2. A new section with comparison to similar software

Great! I have now checked off the final tick mark and can recommend this for publication.
I don't know if it is possible to change the reference style to use numbers instead? See e.g
https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_styles
