### 2020-02-17 26a618a
[JOSS] Update formatting of references
