+++ docs/joss-paper/paper.md
@@ -180,17 +180,17 @@ This not only allows a user to select the most appropriate tool for their resear
 This is an explicit design goal of Chaste, which focusses on the flexibility of implementing multiple models rather than (for example) building a graphical user interface.
 See Table 1 for a comparison of alternatives to Chaste in specific domains, with all other software tools implementing a single modelling framework.
 
-| Software    | Open Source |  GUI  |  CA   |  CP   |  PM   |  VT   |  VM   |             Refs              |
-| ----------- | :---------: | :---: | :---: | :---: | :---: | :---: | :---: | :---------------------------: |
-| Chaste      |      x      |       |   x   |   x   |   x   |   x   |   x   |                               |
-| CompuCell3D |      x      |   x   |       |   x   |       |       |       |    [@Swat2012CompuCell3D]     |
-| Morpheus    |      x      |   x   |       |   x   |       |       |       |    [@Starruss2014Morpheus]    |
-| EPISIM      |             |   x   |       |       |   x   |       |       |    [@Sutterlin2013EPISIM]     |
-| CellSys     |             |   x   |       |       |   x   |       |       |     [@Hoehme2010CellSys]      |
-| PhysiCell   |      x      |       |       |       |   x   |       |       | [@Ghaffarizadeh2018PhysiCell] |
-| Biocellion  |             |       |       |       |   x   |       |       |     [@Kang2014Biocellion]     |
-| VirtualLeaf |      x      |   x   |       |       |       |       |   x   |    [@Merks2011VirtualLeaf]    |
-| EmbryoMaker |      x      |   x   |       |       |   x   |       |       | [@Marin-Riera2015EmbryoMaker] |
+| Software    | Open Source |  GUI  |  CA   |  CP   |  PM   |  VT   |  VM   |
+| ----------- | :---------: | :---: | :---: | :---: | :---: | :---: | :---: |
+| Chaste      |      x      |       |   x   |   x   |   x   |   x   |   x   |
+| CompuCell3D |      x      |   x   |       |   x   |       |       |       |
+| Morpheus    |      x      |   x   |       |   x   |       |       |       |
+| EPISIM      |             |   x   |       |       |   x   |       |       |
+| CellSys     |             |   x   |       |       |   x   |       |       |
+| PhysiCell   |      x      |       |       |       |   x   |       |       |
+| Biocellion  |             |       |       |       |   x   |       |       |
+| VirtualLeaf |      x      |   x   |       |       |       |       |   x   |
+| EmbryoMaker |      x      |   x   |       |       |   x   |       |       |
 
 **Table 1:**
 A comparison of software tools for cell-based modelling.
@@ -200,6 +200,7 @@ CP: cellular Potts.
 PM: particle model, a cell-centre model.
 VT: Voronoi tessellation, a cell-centre model.
 VM: vertex model.
+References: CompuCell3D [@Swat2012CompuCell3D], Morpheus [@Starruss2014Morpheus], EPISIM [@Sutterlin2013EPISIM], CellSys [@Hoehme2010CellSys], PhysiCell [@Ghaffarizadeh2018PhysiCell], Biocellion [@Kang2014Biocellion], VirtualLeaf [@Merks2011VirtualLeaf], EmbryoMaker [@Marin-Riera2015EmbryoMaker].
 
 ### Installation
 
@@ -326,35 +327,35 @@ The initial geometry of major airways segmented from a CT scan (left), and an ex
 Since our last publication on Chaste [@Mirams2013Chaste], over 70 peer-reviewed publications have been enabled, which we mention briefly below.
 
[…]
