### 2020-02-17 #1848 @fcooper8472
@finsberg Great, thanks!

My mistake on the multiple refs - syntax is `;` rather than `,` separation.

For the table, I have just changed it to have the refs below in the caption, which is probably the easiest solution for the time being.
