### 2020-02-24 #1848 @meg-simula
@fcooper8472 Could you take a look at the references missing DOIs as reported by Whedon above?
