### 2020-02-24 54736d5
[JOSS] fix DOIs and formatting
