+++ docs/joss-paper/paper.bib
@@ -128,7 +128,8 @@
   booktitle    = {Annual International Conference of the IEEE Engineering in Medicine and Biology},
   year         = {2010},
   pages        = {760--763},
-  organization = {IEEE}
+  organization = {IEEE},
+  doi          = {10.1109/IEMBS.2010.5626338}
 }
 
 @Article{Bernabeu2010Stimulus,
@@ -169,7 +170,8 @@
   booktitle    = {Annual International Conference of the IEEE Engineering in Medicine and Biology},
   year         = {2010},
   pages        = {6793--6796},
-  organization = {IEEE}
+  organization = {IEEE},
+  doi          = {10.1109/IEMBS.2010.5625979}
 }
 
 @Article{Bordas2011Rabbit,
@@ -200,7 +202,8 @@
   number={12},
   pages={e0144105},
   year={2015},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0144105}
 }
 
 @Article{Britton2013Experimentally,
@@ -233,7 +236,8 @@
   volume={8},
   pages={597},
   year={2017},
-  publisher={Frontiers}
+  publisher={Frontiers},
+  doi={10.3389/fphys.2017.00597}
 }
 
 
@@ -265,7 +269,8 @@
 	number={2},
 	pages={110--126},
 	year={2017},
-	publisher={Taylor \& Francis}
+	publisher={Taylor \& Francis},
+  doi={10.1080/21681163.2015.1023358}
 }
 
 @Article{Campos2016Lattice,
@@ -306,7 +311,8 @@
 	number={suppl\_4},
 	pages={iv4--iv15},
 	year={2016},
-	publisher={Oxford University Press}
+	publisher={Oxford University Press},
+  doi={10.1093/europace/euw346}
 }
 
 @article{Cardone2018,
@@ -465,7 +471,8 @@
   title     = {A novel biophysically-detailed mathematical model of rabbit Purkinje cell electrophysiology.},
   booktitle = {Annual International Conference of the IEEE Engineering in Medicine and Biology Society},
   year      = {2010},
-  pages     = {2658}
+  pages     = {2658},
+  doi       = {10.1109/IEMBS.2010.5626614}
 }
 
 @Article{Corrias2011Ionic,
@@ -518,7 +525,7 @@
   volume  = {14},
   number  = {134},
   pages   = {20170340},
-  doi     = {1098/rsif.2017.0340}
+  doi     = {10.1098/rsif.2017.0340}
 }
 
 @article{Daly2018,
@@ -683,7 +690,8 @@
   journal = {Multiscale Cancer Modeling, Editor: TS Deisboeck},
   year    = {2010},
   volume  = {6},
-  pages   = {111--134}
+  pages   = {111--134},
+  doi     = {10.1201/b10407-7}
 }
 
 @Article{Fletcher2011Novel,
@@ -744,8 +752,10 @@
   author  = {Fletcher, A.G. and Cooper, F. and Baker, R.E.},
   title   = {{Mechanocellular models of epithelial morphogenesis}},
   journal = {Phil Trans R Soc B},
-  year    = {2016},
-  volume  = {In press}
+  volume  = {372},
+  number  = {1720},
+  year    = {2017},
+  doi     = {10.1098/rstb.2015.0519},
 }
 
 @InProceedings{Franzetti2015Combined,
@@ -811,14 +821,14 @@
 }
 
 @Article{Gavaghan2012Bidomain,
-  author  = {Gavaghan, D. and Rodriguez, B.},
-  title   = {{A bidomain model of the cardiac specialized conduction system of the heart}},
[…]
