### 2020-02-24 #1848 @fcooper8472
@meg-simula I have fixed the DOIs and the formatting issue.

No idea why whedon can't generate a new pdf - I only changed a single line in the `paper.md` - it's definitely still there!  (And the references were found and checked fine!)
