### 2020-02-25 #1848 @meg-simula
@fcooper8472 Looks like there is still a few references missing dois/link for instance some of the Spiteri papers - could you take another look? (Sorry to keep niggling about this.)
