### 2020-02-26 412a28e
[JOSS] more tweaking of refs
