+++ docs/joss-paper/paper.bib
@@ -11,7 +11,7 @@
 @Article{Abramson2010High,
   author  = {Abramson, D. and Bernabeu, M.O. and Bethwaite, B. and Burrage, K. and Corrias, A. and Enticott, C. and Garic, S. and Gavaghan, D. and Peachey, T. and Pitt-Francis, J. and Pueyo, E. and Rodriguez, B. and Sher, A. and Tan, J.},
   title   = {{High-throughput cardiac science on the Grid}},
-  journal = {Phil Trans R Soc A},
+  journal = {Philos. Trans. R. Soc. A},
   year    = {2010},
   volume  = {368},
   number  = {1925},
@@ -22,7 +22,7 @@
 @Article{Agudelo-Toro2013Computationally,
   author  = {Agudelo-Toro, A. and Neef, A.},
   title   = {{Computationally efficient simulation of electrical activity at cell membranes interacting with self-generated and externally imposed electric fields}},
-  journal = {J Neural Eng},
+  journal = {J. Neural Eng.},
   year    = {2013},
   volume  = {10},
   number  = {2},
@@ -33,7 +33,7 @@
 @Article{Almet2018Multicellular,
   author  = {Almet, A.A. and Hughes, B.D. and Landman, K.A. and N{\"a}thke, I.S. and Osborne, J.M.},
   title   = {{A multicellular model of intestinal crypt buckling and fission}},
-  journal = {Bull Math Biol},
+  journal = {Bull. Math. Biol.},
   year    = {2018},
   volume  = {80},
   number  = {2},
@@ -44,7 +44,7 @@
 @Article{Arthurs2012Efficient,
   author  = {Arthurs, C.J. and Bishop, M.J. and Kay, D.},
   title   = {{Efficient simulation of cardiac electrical propagation using high order finite elements}},
-  journal = {J Comput Phys},
+  journal = {J. Comput. Phys},
   year    = {2012},
   volume  = {231},
   number  = {10},
@@ -106,7 +106,7 @@
 @Article{Bernabeu2009Chaste,
   author  = {Bernabeu, M. and Bordas, R. and Pathmanathan, P. and Pitt-Francis, J. and Cooper, J. and Garny, A. and Gavaghan, D.J. and Rodr{\'i}guez, B. and Southern, J.A. and Whiteley, J.P.},
   title   = {{Chaste: incorporating a novel multi-scale spatial and temporal algorithm into a large-scale open source library}},
-  journal = {Phil Trans R Soc A},
+  journal = {Philos. Trans. R. Soc. A},
   year    = {2009},
   volume  = {367},
   pages   = {1907-1930},
@@ -187,7 +187,7 @@
 @Article{Bordas2012Bidomain,
   author  = {Bordas, R. and Gillow, K. and Gavaghan, D. and Rodriguez, B. and Kay, D.},
   title   = {{A bidomain model of the ventricular specialized conduction system of the heart}},
-  journal = {SIAM J Appl Math},
+  journal = {SIAM J. Appl. Math},
   year    = {2012},
   volume  = {72},
   pages   = {1618--1643},
@@ -347,12 +347,13 @@
 @article{Cervi2018HighOrder,
   title={High-order operator splitting for the bidomain and monodomain models},
   author={Cervi, Jessica and Spiteri, Raymond J},
-  journal={SIAM Journal on Scientific Computing},
+  journal={SIAM J. Sci. Comput},
   volume={40},
   number={2},
   pages={A769--A786},
   year={2018},
-  publisher={SIAM}
+  publisher={SIAM},
[…]
