### 2020-02-26 #1848 @fcooper8472
@meg-simula There are now DOIs for all references other than the Computing in Cardiology articles, which don't appear to have DOIs.
