### 2020-02-28 794a07f
Edge split due division is recorded as pre remap event.
New bug found: edges dissappear or generated at wrong places. This commit contains outputs to debug. In plans: Death/division/remesh leads to rebuilding of the edges
