+++ cell_based/src/simulation/AbstractCellBasedSimulation.cpp
@@ -37,7 +37,7 @@ OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <iostream>
 #include <fstream>
 #include <set>
-
+#include <chrono>
 #include "AbstractCellBasedSimulation.hpp"
 #include "CellBasedEventHandler.hpp"
 #include "LogFile.hpp"
@@ -429,8 +429,19 @@ void AbstractCellBasedSimulation<ELEMENT_DIM,SPACE_DIM>::Solve()
     {
         LOG(1, "--TIME = " << p_simulation_time->GetTime() << "\n");
 
+        std::clock_t c_start_total = std::clock();
+        auto t_start_total = std::chrono::high_resolution_clock::now();
+
+        std::clock_t c_start_remesh = std::clock();
+        auto t_start_remesh = std::chrono::high_resolution_clock::now();
+
         // This function calls DoCellRemoval(), DoCellBirth() and CellPopulation::Update()
         UpdateCellPopulation();
+
+        std::clock_t c_end_remesh = std::clock();
+        auto t_end_remesh = std::chrono::high_resolution_clock::now();
+        const double cpu_time_remesh = (c_end_remesh-c_start_remesh)/CLOCKS_PER_SEC;
+        const double wall_time_remesh = std::chrono::duration_cast<std::chrono::nanoseconds>(t_end_remesh - t_start_remesh).count();
 
         // Store whether we are sampling results at the current timestep
         SimulationTime* p_time = SimulationTime::Instance();
@@ -458,16 +469,30 @@ void AbstractCellBasedSimulation<ELEMENT_DIM,SPACE_DIM>::Solve()
          * Note that UpdateCellPopulation() may affect cell properties that influence cell locations
          * and topology update. Therefore, simulation modifiers must be called before UpdateCellLocationsAndTopology()*/
         CellBasedEventHandler::BeginEvent(CellBasedEventHandler::UPDATESIMULATION);
+        std::clock_t c_start_update = std::clock();
+        auto t_start_update = std::chrono::high_resolution_clock::now();
         for (typename std::vector<boost::shared_ptr<AbstractCellBasedSimulationModifier<ELEMENT_DIM, SPACE_DIM> > >::iterator iter = mSimulationModifiers.begin();
                 iter != mSimulationModifiers.end();
                 ++iter)
         {
             (*iter)->UpdateAtEndOfTimeStep(this->mrCellPopulation);
         }
+        std::clock_t c_end_update = std::clock();
+        auto t_end_update = std::chrono::high_resolution_clock::now();
+        const double cpu_time_update = (c_end_update-c_start_update)/CLOCKS_PER_SEC;
+        const double wall_time_update = std::chrono::duration_cast<std::chrono::nanoseconds>(t_end_update - t_start_update).count();
         CellBasedEventHandler::EndEvent(CellBasedEventHandler::UPDATESIMULATION);
+
+        std::clock_t c_start_nodes = std::clock();
+        auto t_start_nodes = std::chrono::high_resolution_clock::now();
 
         // Update cell locations and topology
         UpdateCellLocationsAndTopology();
+
+        std::clock_t c_end_nodes = std::clock();
+        auto t_end_nodes = std::chrono::high_resolution_clock::now();
+        const double cpu_time_nodes = (c_end_nodes-c_start_nodes)/CLOCKS_PER_SEC;
+        const double wall_time_nodes = std::chrono::duration_cast<std::chrono::nanoseconds>(t_end_nodes - t_start_nodes).count();
 
         // Now write cell velocities to file if required
[…]
