+++ cell_based/src/population/srn_update/VertexBasedPopulationSrn.cpp
@@ -49,53 +49,71 @@ void VertexBasedPopulationSrn<DIM>::UpdateSrnAfterBirthOrDeath(VertexElementMap&
 {
     std::vector<EdgeOperation*> edge_operations
     = mpCellPopulation->rGetMesh().GetOperationRecorder()->GetEdgeOperations();
+    /*std::cout<<"OP set: "<<std::endl;
+    for (unsigned int i=0; i<edge_operations.size(); ++i)
+        std::cout<<" "<<edge_operations[i]->GetOperation();
+    std::cout<<std::endl;*/
     for (auto operation:edge_operations)
     {
+        //std::cout<<"OP: "<<operation->GetOperation()<<std::endl;
         // An operation with deleted element may be recorded,
         // e.g. following a T2 swap, neighbouring element may become triangular due to node merging
-        // and may subsequently be marked for deletion, despite being marked to perform edge operations.
-        if (!rElementMap.IsDeleted(operation->GetElementIndex()))
+        // and may subsequently be marked for deletion, despite being marked to perform edge operations (e.g. node merging).
+
+        switch (operation->GetOperation())
         {
-            switch (operation->GetOperation())
-            {
-            case EDGE_OPERATION_ADD:
-            case EDGE_OPERATION_NODE_MERGE:
-            case EDGE_OPERATION_SPLIT:
-            case EDGE_OPERATION_MERGE:
-            {
-                unsigned int location_index = operation->GetElementIndex();
-                if (operation->IsElementIndexRemapped())
-                {
-                    location_index = rElementMap.GetNewIndex(operation->GetElementIndex());
-                }
-                EdgeRemapInfo *pEdgeChange = operation->GetRemapInfo();
-                CellPtr cell = mpCellPopulation->GetCellUsingLocationIndex(location_index);
-                auto old_model = static_cast<SrnCellModel*>(cell->GetSrnModel());
-                std::vector<AbstractSrnModelPtr> old_srn_edges = old_model->GetEdges();
-                RemapCellSrn(old_srn_edges, old_model, pEdgeChange);
-                old_srn_edges.clear();
-                break;
-            }
-            case EDGE_OPERATION_DIVIDE:
-            {
-                const unsigned int location_index_1 = rElementMap.GetNewIndex(operation->GetElementIndex());
-                const unsigned int location_index_2 = rElementMap.GetNewIndex(operation->GetElementIndex2());
-                EdgeRemapInfo *pEdgeChange_1 = operation->GetRemapInfo();
-                EdgeRemapInfo *pEdgeChange_2 = operation->GetRemapInfo2();
-                CellPtr cell_1 = mpCellPopulation->GetCellUsingLocationIndex(location_index_1);
-                CellPtr cell_2 = mpCellPopulation->GetCellUsingLocationIndex(location_index_2);
+        case EDGE_OPERATION_ADD:
+        case EDGE_OPERATION_NODE_MERGE:
+        case EDGE_OPERATION_SPLIT:
+        case EDGE_OPERATION_MERGE:
+        {
+            const unsigned int stored_index = operation->GetElementIndex();
+            unsigned int location_index = stored_index;
 
-                auto old_model_1 = static_cast<SrnCellModel*>(cell_1->GetSrnModel());
-                std::vector<AbstractSrnModelPtr> parent_srn_edges = old_model_1->GetEdges();
-                auto old_model_2 = static_cast<SrnCellModel*>(cell_2->GetSrnModel());
[…]
