### 2020-03-09 07eb4d7
Element edges are rebuilt after swaps
Removed comments for debugging too
