+++ mesh/src/vertex/MutableVertexMesh.cpp
@@ -361,7 +361,6 @@ unsigned MutableVertexMesh<ELEMENT_DIM, SPACE_DIM>::DivideElementAlongGivenAxis(
 
     // Get the centroid of the element
     c_vector<double, SPACE_DIM> centroid = this->GetCentroidOfElement(pElement->GetIndex());
-    std::cout<<"Dividing cell loc: "<<centroid(0)<<" "<<centroid(1)<<std::endl;
     // Create a vector perpendicular to the axis of division
     c_vector<double, SPACE_DIM> perp_axis;
     perp_axis(0) = -axisOfDivision(1);
@@ -543,8 +542,6 @@ unsigned MutableVertexMesh<ELEMENT_DIM, SPACE_DIM>::DivideElementAlongGivenAxis(
                 double theta = prev_curr_distance/old_distance;
                 relative_new_node.push_back(theta);
                 edge_split_pairs.push_back(std::pair<VertexElement<ELEMENT_DIM,SPACE_DIM>*, unsigned int>(p_element,index));
-                auto elem_loc = this->GetCentroidOfElement(p_element->GetIndex());
-                std::cout<<"Other split el location: "<<p_element->GetIndex()<<" "<<elem_loc(0)<<" "<<elem_loc(1)<<std::endl;
             }
             // Add new node to this element
             p_element->AddNode(this->GetNode(new_node_global_index), index);
@@ -1493,6 +1490,8 @@ void MutableVertexMesh<ELEMENT_DIM, SPACE_DIM>::PerformNodeMerge(Node<SPACE_DIM>
 
     // Update the elements previously containing node B to contain node A
     unsigned node_B_index = pNodeB->GetIndex();
+    // For rebuilding edges affected elements
+    std::set<unsigned> rebuilt_elements;
     for (std::set<unsigned>::const_iterator it = nodeB_elem_indices.begin(); it != nodeB_elem_indices.end(); ++it)
     {
         // Find the local index of node B in this element
@@ -1524,11 +1523,14 @@ void MutableVertexMesh<ELEMENT_DIM, SPACE_DIM>::PerformNodeMerge(Node<SPACE_DIM>
             // Replace node B with node A in this element
             this->mElements[*it]->UpdateNode(node_B_local_index, pNodeA);
         }
+        rebuilt_elements.insert(this->mElements[*it]->GetIndex());
     }
 
     assert(!(this->mNodes[node_B_index]->IsDeleted()));
     this->mNodes[node_B_index]->MarkAsDeleted();
     mDeletedNodeIndices.push_back(node_B_index);
+    for (unsigned i:rebuilt_elements)
+        this->GetElement(i)->RebuildEdges();
 }
 
 template<unsigned ELEMENT_DIM, unsigned SPACE_DIM>
@@ -1600,6 +1602,8 @@ void MutableVertexMesh<ELEMENT_DIM, SPACE_DIM>::PerformT1Swap(Node<SPACE_DIM>* p
     std::set<unsigned> nodeA_elem_indices = pNodeA->rGetContainingElementIndices();
     std::set<unsigned> nodeB_elem_indices = pNodeB->rGetContainingElementIndices();
 
+    //For rebuilding affected elements
+    std::set<unsigned> rebuilt_elements;
     for (std::set<unsigned>::const_iterator it = rElementsContainingNodes.begin();
          it != rElementsContainingNodes.end();
          ++it)
@@ -1625,6 +1629,7 @@ void MutableVertexMesh<ELEMENT_DIM, SPACE_DIM>::PerformT1Swap(Node<SPACE_DIM>* p
             this->mElements[*it]->AddNode(pNodeA, nodeB_local_index);
             if (mTrackMeshOperations)
[…]
