### 2020-03-10 a1eb100
[JOSS] correct affiliation
