+++ docs/joss-paper/paper.md
@@ -89,7 +89,7 @@ authors:
    orcid: 0000-0002-3435-7388
    affiliation: 3
  - name: Raymond J Spiteri
-   affiliation: 8
+   affiliation: 10
  - name: David J Gavaghan
    orcid: 0000-0001-8311-3200
    affiliation: 3
