+++ docs/joss-paper/paper.bib
@@ -218,15 +218,15 @@
 }
 
 @ARTICLE{Britton2017,
-	AUTHOR={Britton, Oliver J. and Bueno-Orovio, Alfonso and Virág, László and Varró, András and Rodriguez, Blanca},
-	TITLE={The Electrogenic Na+/K+ Pump Is a Key Determinant of Repolarization Abnormality Susceptibility in Human Ventricular Cardiomyocytes: A Population-Based Simulation Study},
-	JOURNAL={Frontiers in Physiology},
-	VOLUME={8},
-	PAGES={278},
-	YEAR={2017},
-	URL={https://www.frontiersin.org/article/10.3389/fphys.2017.00278},
-	DOI={10.3389/fphys.2017.00278},
-	ISSN={1664-042X},
+  AUTHOR={Britton, Oliver J. and Bueno-Orovio, Alfonso and Virág, László and Varró, András and Rodriguez, Blanca},
+  TITLE={{The electrogenic Na+/K+ pump is a key determinant of repolarization abnormality susceptibility in human ventricular cardiomyocytes: A population-based simulation study}},
+  JOURNAL={Frontiers in Physiology},
+  VOLUME={8},
+  PAGES={278},
+  YEAR={2017},
+  URL={https://www.frontiersin.org/article/10.3389/fphys.2017.00278},
+  DOI={10.3389/fphys.2017.00278},
+  ISSN={1664-042X},
 }
 
 @article{britton2017quantitative,
@@ -262,14 +262,14 @@
 }
 
 @article{burrowes2017combined,
-	title={A combined image-modelling approach assessing the impact of hyperinflation due to emphysema on regional ventilation--perfusion matching},
-	author={Burrowes, KS and Doel, Tom and Kim, M and Vargas, Claudia and Roca, Josep and Grau, Vicente and Kay, D},
-	journal={Computer Methods in Biomechanics and Biomedical Engineering: Imaging \& Visualization},
-	volume={5},
-	number={2},
-	pages={110--126},
-	year={2017},
-	publisher={Taylor \& Francis},
+  title={A combined image-modelling approach assessing the impact of hyperinflation due to emphysema on regional ventilation--perfusion matching},
+  author={Burrowes, KS and Doel, Tom and Kim, M and Vargas, Claudia and Roca, Josep and Grau, Vicente and Kay, D},
+  journal={Computer Methods in Biomechanics and Biomedical Engineering: Imaging \& Visualization},
+  volume={5},
+  number={2},
+  pages={110--126},
+  year={2017},
+  publisher={Taylor \& Francis},
   doi={10.1080/21681163.2015.1023358}
 }
 
@@ -304,32 +304,32 @@
 }
 
 @article{cardone2016human,
-	title={Human ventricular activation sequence and the simulation of the electrocardiographic {QRS} complex and its variability in healthy and intraventricular block conditions},
-	author={Cardone-Noott, L. and Bueno-Orovio, A. and Minchole, A. and Zemzemi, N. and Rodriguez, B.},
-	journal={EP Europace},
-	volume={18},
-	number={suppl\_4},
-	pages={iv4--iv15},
-	year={2016},
-	publisher={Oxford University Press},
+  title={Human ventricular activation sequence and the simulation of the electrocardiographic {QRS} complex and its variability in healthy and intraventricular block conditions},
+  author={Cardone-Noott, L. and Bueno-Orovio, A. and Minchole, A. and Zemzemi, N. and Rodriguez, B.},
+  journal={EP Europace},
+  volume={18},
+  number={suppl\_4},
[…]
