### 2020-03-10 #1848 @fcooper8472
Thanks @meg-simula

When checking the Zenodo metadata @mirams found one of the author affiliations was incorrect so I have made a new commit fixing that.  It might be worth checking the proof again just in case something has gone wrong.

In terms of the other information you requested:

>  Make a tagged release of your software, and list the version tag of the archived version here.

Tagged as `release_2019.1` ([link](https://github.com/Chaste/Chaste/releases/tag/release_2019.1))

> Archive the reviewed software in Zenodo or a similar service

Done: https://zenodo.org/record/3702736

> Please list the DOI of the archived version here.

[10.5281/zenodo.3702736](https://doi.org/10.5281/zenodo.3702736)

The Zenodo metadata will be updated shortly.
