### 2020-03-10 #1848 @danielskatz
👋 @fcooper8472 - as the AEiC currently on duty, I have some requests before this goes into the final acceptance processing

- [ ] please remove a large number of the reference related to use of Chaste -  you do not need to cite all papers in which Chaste has been used, just a selection showing its breadth, perhaps 5-10.  This paper is far longer than a JOSS paper is recommended to be, and while some of this is because it is a complex package, we can reduce some of the length and number of references by doing this.

- [ ] please fix paper and journal titles - e.g.,  I see both PLoS one and PLOS ONE, where the latter is correct, and Computing in cardiology, rather than Computing in Cardiology.  Also titles, I see one lgr+ that should be Lgr5+, for example.
