### 2020-03-11 ababb7e
[JOSS] update bib file
