+++ docs/joss-paper/paper.bib
@@ -66,7 +66,7 @@
 @Article{Baker2014Quantification,
   author  = {Baker, A.-M. and Cereser, B. and Melton, S. and Fletcher, A.G. and Rodriguez-Justo, M. and Tadrous, P.J. and Humphries, A. and Elia, G. and McDonald, S.A.C. and Wright, N.A. and Simons, B.D. and Jansen, M. and Graham, T.A.},
   title   = {{Quantification of crypt and stem cell evolution in the normal and neoplastic human colon}},
-  journal = {Cell Rep},
+  journal = {Cell Rep.},
   year    = {2014},
   volume  = {8},
   number  = {4},
@@ -77,7 +77,7 @@
 @InProceedings{Bartolucci2014Linking,
   author       = {Bartolucci, C. and Moreno, C. and de la Cruz, A. and Lambiase, P. and Severi, S. and Valenzuela, C.},
   title        = {{Linking a novel mutation to its short QT phenotype through multiscale computational modelling}},
-  booktitle    = {Computing in Cardiology 2014},
+  booktitle    = {{Computing in Cardiology}},
   year         = {2014},
   pages        = {1017--1020},
   organization = {IEEE}
@@ -86,7 +86,7 @@
 @Article{Beattie2013Evaluation,
   author  = {Beattie, K.A. and Luscombe, C. and Williams, G. and Munoz-Muriedas, J. and Gavaghan, D.J. and Cui, Y. and Mirams, G.R.},
   title   = {{Evaluation of an in silico cardiac safety assay: using ion channel screening data to predict QT interval changes in the rabbit ventricular wedge}},
-  journal = {J Pharmacol Toxicol Methods},
+  journal = {J. Pharmacol. Toxicol. Methods},
   year    = {2013},
   volume  = {68},
   number  = {1},
@@ -97,7 +97,7 @@
 @InProceedings{Bernabeu2008High,
   author    = {Bernabeu, M.O. and Bishop, M.J. and Pitt-Francis, J. and Gavaghan, D.J. and Grau, V. and Rodriguez, B.},
   title     = {{High performance computer simulations for the study of biological function in 3D heart models incorporating fibre orientation and realistic geometry at para-cellular resolution}},
-  booktitle = {Computers in Cardiology},
+  booktitle = {{Computers in Cardiology}},
   year      = {2008},
   pages     = {721--724},
   doi       = {10.1109/CIC.2008.4749143}
@@ -116,7 +116,7 @@
 @Article{Bernabeu2009Grid,
   author  = {Bernabeu, M.O. and Corrias, A. and Pitt-Francis, J. and Rodriguez, B. and Bethwaite, B. and Enticott, C. and Garic, S. and Peachey, T. and Tan, J. and Abramson, D. and Gavaghan, D.},
   title   = {{Grid computing simulations of ion channel block effects on the ECG using 3D anatomically-based models}},
-  journal = {Comput Cardiol},
+  journal = {Comput. Cardiol.},
   year    = {2009},
   volume  = {36},
   pages   = {213--216}
@@ -124,8 +124,8 @@
 
 @InProceedings{Bernabeu2010Shock,
   author       = {Bernabeu, M.O. and Wallman, M. and Rodriguez, B.},
-  title        = {Shock-induced arrhythmogenesis in the human heart: A computational modelling study},
-  booktitle    = {Annual International Conference of the IEEE Engineering in Medicine and Biology},
+  title        = {Shock-induced arrhythmogenesis in the human heart: a computational modelling study},
+  booktitle    = {{Annual International Conference of the IEEE Engineering in Medicine and Biology Society (EMBC)}},
[…]
