### 2020-03-11 #1848 @fcooper8472
@danielskatz, please see responses inline:

> please remove a large number of the reference related to use of Chaste - you do not need to cite all papers in which Chaste has been used, just a selection showing its breadth, perhaps 5-10.

We added a comprehensive list of research using/enabled by Chaste because that was listed on the JOSS website under "What should my paper contain?", which included:

"Mention (if applicable) of any past or ongoing research projects using the software and recent scholarly publications enabled by it."

The paper is longer than the recommended limit of 1000 words because that seemed necessary in order to hit all of the submission requirements.  I notice that, since yesterday, the JOSS website has changed, and the new instructions are to:

"Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it."

In light of the fact that we submitted our paper to JOSS before the requirements changed, it doesn't seem appropriate to make substantial changes at this stage.

> please fix paper and journal titles - e.g., I see both PLoS one and PLOS ONE, where the latter is correct, and Computing in cardiology, rather than Computing in Cardiology. Also re titles, I see one lgr+ that should be Lgr5+, for example.

I have gone thoroughly through the bibliography file and checked journal names and titles.  They should now be correct.

The issue with Computing in cardiology, rather than Computing in Cardiology appears to be a JOSS problem - all instances of Cardiology in the bibliography file were already capitalised, and they became lowercase when converted to PDF.  I have tried to circumvent this by wrapping in an extra layer of `{}` which will hopefully mean they are displayed verbatim.
