### 2020-03-11 #1848 @danielskatz
> In light of the fact that we submitted our paper to JOSS before the requirements changed, it doesn't seem appropriate to make substantial changes at this stage.

I disagree - please make the changes I requested.

> I have gone thoroughly through the bibliography file and checked journal names and titles. They should now be correct.

Thank you

> The issue with Computing in cardiology, rather than Computing in Cardiology appears to be a JOSS problem - all instances of Cardiology in the bibliography file were already capitalised, and they became lowercase when converted to PDF. I have tried to circumvent this by wrapping in an extra layer of {} which will hopefully mean they are displayed verbatim.

Rather than a problem with JOSS, we could call this a feature that comes from pandoc - titles, including booktitles, are changed to only an initial capital letter.
