### 2020-03-12 #1848 @mirams
Hi, I am one of the authors of the Chaste paper, I have spent a while trying to shorten it down as requested. I would ask that you take another look, for a couple of reasons.

Firstly in my draft I've had to remove all the citations to similar software that feature in the comparison table we added (as requested before), this seems against the spirit of JOSS in giving credit to these people for developing their software.

Secondly, the section on "Recent Publications..." gives a very good indication to readers of what Chaste can be used for, and where they can find more information on this, which is a far wider range of things than the limited examples in the paper can hope to cover.

If you would still like to chop it down we're happy to oblige, but personally I think it makes the paper less useful.
