### 2020-03-12 #1848 @danielskatz
The original request I made was "you do not need to cite all papers in which Chaste has been used, just a selection showing its breadth, perhaps 5-10." - I did not mean that you should remove references to things that Chaste depends on, or those to similar software, just to reduce the number of references in the section "Recent Publications enabled by Chaste" to a representative 5 to 10 rather than every paper that has used Chaste. You can still say that Chaste can be used in a variety of fields without referencing papers from each of them.

I apologize if this was confusing.

Also, when you make changes, you can ask whedon to create a new PDF by putting the command ```@whedon generate pdf``` in this issue as a new comment
