### 2020-03-12 cdc56f3
[JOSS] Remove most references
