+++ docs/joss-paper/paper.md
@@ -325,38 +325,30 @@ The initial geometry of major airways segmented from a CT scan (left), and an ex
 
 ## Recent publications enabled by Chaste
 
-Since our last publication on Chaste [@Mirams2013Chaste], over 70 peer-reviewed publications have been enabled, which we mention briefly below.
+Since our last publication on Chaste, over 70 peer-reviewed publications have been enabled in the areas described below.
 
 Publications using Cardiac Chaste have included scientific studies relating to:
-basic mechanisms of cardiac electrophysiology in healthy and diseased settings [@Walmsley2013mRNA; @Passini2013Computational; @Sadrieh2014Multiscale; @Bartolucci2014Linking; @Samanta2015Ca; @Pathmanathan2015; @cardone2016human; @Zhou2016; @Mahoney2016Connexin43; @Corsi2017Noninvasive; @Dutta2016; @Reilly2016; @Dutta2017; @Lyon2018;  @Xin2019];
-the effects of realistic tissue structure on simulated cardiac electrical activity [@Walmsley2013Estimation; @Lekadir2014Effect; @Lekadir2016Statistically; @Zacur2017];
-the sources and consequences of inter-subject electrophysiological variability [@Dutta2013Ionic; @Britton2013Experimentally; @Elkins2013Variability; @Walmsley2015Application; @Britton2017; @Muszkiewicz2018];
-predicting the effects of drugs on cardiac activity, including safety assessment [@Beattie2013Evaluation; @Zemzemi2013Computational; @Wallman2014Computational; @Mirams2014Prediction; @Passini2014Late; @Cardone-Noott2014Computational; @Zemzemi2015Effects; @Moreno2015New; @Davies2016Recent; @Hill2016Computational; @Passini2016; @britton2017quantitative; @McMillan2017; @passini2017human; @Lim2018Role];
-and the development of associated web-based tools [@Williams2015Web; @Cooper2016Cardiac; @Daly2018].
-Other studies enabled by Cardiac Chaste have advanced the
-methodologies for parameter identifiability and inference, model
-selection and uncertainty quantification for in cardiac
-electrophysiology models
-[@Daly2015Hodgkin; @Mirams2016White; @Johnstone2016Uncertainty; @Daly2017Comparing];
-and for the verification and efficient numerical simulation of such
-models
-[@Marsh2012Secrets;@Agudelo-Toro2013Computationally; @Pathmanathan2014Verification; @Corrado2016Stability; @Campos2016Lattice; @Spiteri2016Godunov; @Cervi2018HighOrder; @Cardone2018; @Green2019LUT].
-The continuum-mechanics solvers in Chaste have been used for studies of dielectric elastomers [@Langham2018Modeling]; and our electro-mechanics code as used in [@Carapella2014Quantitative], has also been used to verify new numerical methods [@gurev2015high].
+basic mechanisms of cardiac electrophysiology and tissue structure in healthy and diseased settings;
+the sources and consequences of inter-subject electrophysiological variability;
+predicting the effects of drugs on cardiac activity, including safety assessment
+and the development of associated web-based tools [@Cooper2016Cardiac].
+Other studies enabled by Cardiac Chaste have advanced technical methodologies for parameter identifiability and inference, model selection and uncertainty quantification in cardiac electrophysiology models [@Johnstone2016Uncertainty];
+and for the verification and efficient numerical simulation of cardiac
+models [@Green2019LUT].
+The continuum-mechanics solvers in Chaste have been used for studies of cardiac electromechanics [@Carapella2014Quantitative].
 
[…]
