+++ docs/joss-paper/paper.bib
@@ -196,7 +196,7 @@
 
 @article{bordas2015development,
   title={Development and analysis of patient-based complete conducting airways models},
-  author={Bordas, Rafel and Lefevre, Christophe and Veeckmans, Bart and Pitt-Francis, Joe and Fetita, Catalin and Brightling, Christopher E and Kay, David and Siddiqui, Salman and Burrowes, Kelly S},
+  author={Bordas, R. and Lefevre, C. and Veeckmans, B. and Pitt-Francis, J. and Fetita, C. and Brightling, C.E. and Kay, D. and Siddiqui, S. and Burrowes, K.S.},
   journal={PlOS ONE},
   volume={10},
   number={12},
@@ -231,7 +231,7 @@
 
 @article{britton2017quantitative,
   title={Quantitative comparison of effects of dofetilide, sotalol, quinidine, and verapamil between human ex vivo trabeculae and in silico ventricular models incorporating inter-individual action potential variability},
-  author={Britton, Oliver J and Abi-Gerges, Najah and Page, Guy and Ghetti, Andre and Miller, Paul E and Rodriguez, Blanca},
+  author={Britton, Oliver J. and Abi-Gerges, Najah and Page, Guy and Ghetti, Andre and Miller, Paul E and Rodriguez, Blanca},
   journal={Front. Physiol.},
   volume={8},
   pages={597},
@@ -1374,7 +1374,7 @@ doi     = {10.1016/j.bpj.2013.11.4498}
 
 @article{passini2017human,
   title={Human in silico drug trials demonstrate higher accuracy than animal models in predicting clinical pro-arrhythmic cardiotoxicity},
-  author={Passini, Elisa and Britton, Oliver J and Lu, Hua Rong and Rohrbacher, Jutta and Hermans, An N and Gallacher, David J and Greig, Robert JH and Bueno-Orovio, Alfonso and Rodriguez, Blanca},
+  author={Passini, Elisa and Britton, Oliver J. and Lu, Hua Rong and Rohrbacher, Jutta and Hermans, An N and Gallacher, David J and Greig, Robert JH and Bueno-Orovio, Alfonso and Rodriguez, Blanca},
   journal={Front. Physiol.},
   volume={8},
   pages={668},
