### 2020-03-12 #1848 @fcooper8472
@danielskatz here is the substantially lighter version.

Please let me know if you would like any further changes.  We will need to generate a new archive & DOI reflecting these changes before publication.
