### 2020-03-12 #1848 @fcooper8472
The new Zenodo archive is:
https://zenodo.org/record/3708497

with DOI:
[10.5281/zenodo.3708497](http://doi.org/10.5281/zenodo.3708497)

@mirams will be able to update the title and author list at that new Zenodo archive to match this JOSS paper.
